Adds support for zone offset adjustment.
The purpose behind zone offset adjustment is to be able to edit a time
representing something in another time zone, even though the timestamps
involved are still milliseconds from Epoch.

Note: DST boundary handling is a mess larger than the scope of this plugin.
This is because JavaScript's Date() object is really, really simple.
The fact is that Date() is bound to the rules of your system's DST.  So
if you're in EST trying to edit a time local to Arizona (which has its
own DST rules), you're going to notice the times not following the
Arizona rules.

The good news is that $(element).timeInput('value') returns a timestamp,
and that timestamp will be correct.  You can check it against the tz
database if you need to do something special with it.  The bad news is that this
plugin will still malform the visual representation of it.  If you need
to use this plugin and DST boundaries are an issue, I'd recommend
binding to $(element).change() or $(element).blur() to look it up using
ajax and return the "proper" string representation.